Revert "Feature/rearranging code files"
