--- .env.example
@@ -1,2 +1,0 @@
-AUTH_EMAIL=
-AUTH_PASS=
