--- controllers/DateController.js
@@ -1,9 +1,0 @@
-const currentDateTime = (value) => {
-    
-    return (Date.parse(value))
-  }
-
-module.exports = {
-    currentDateTime:currentDateTime
-}
-
